// FamilyClub.React/famiy.club.react/src/lib/api/generated/models/OrderDTO.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FamilyClub.WebAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { OrderItemDTO } from './OrderItemDTO';
import {
    OrderItemDTOFromJSON,
    OrderItemDTOFromJSONTyped,
    OrderItemDTOToJSON,
    OrderItemDTOToJSONTyped,
} from './OrderItemDTO';

/**
 * 
 * @export
 * @interface OrderDTO
 */
export interface OrderDTO {
    /**
     * 
     * @type {number}
     * @memberof OrderDTO
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof OrderDTO
     */
    userId?: number;
    /**
     * 
     * @type {Date}
     * @memberof OrderDTO
     */
    orderDate?: Date;
    /**
     * 
     * @type {string}
     * @memberof OrderDTO
     */
    status?: string | null;
    /**
     * 
     * @type {number}
     * @memberof OrderDTO
     */
    totalPrice?: number;
    /**
     * 
     * @type {Array<OrderItemDTO>}
     * @memberof OrderDTO
     */
    orderItems?: Array<OrderItemDTO> | null;
}

/**
 * Check if a given object implements the OrderDTO interface.
 */
export function instanceOfOrderDTO(value: object): value is OrderDTO {
    return true;
}

export function OrderDTOFromJSON(json: any): OrderDTO {
    return OrderDTOFromJSONTyped(json, false);
}

export function OrderDTOFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderDTO {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'userId': json['userId'] == null ? undefined : json['userId'],
        'orderDate': json['orderDate'] == null ? undefined : (new Date(json['orderDate'])),
        'status': json['status'] == null ? undefined : json['status'],
        'totalPrice': json['totalPrice'] == null ? undefined : json['totalPrice'],
        'orderItems': json['orderItems'] == null ? undefined : ((json['orderItems'] as Array<any>).map(OrderItemDTOFromJSON)),
    };
}

export function OrderDTOToJSON(json: any): OrderDTO {
    return OrderDTOToJSONTyped(json, false);
}

export function OrderDTOToJSONTyped(value?: OrderDTO | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'userId': value['userId'],
        'orderDate': value['orderDate'] == null ? value['orderDate'] : value['orderDate'].toISOString(),
        'status': value['status'],
        'totalPrice': value['totalPrice'],
        'orderItems': value['orderItems'] == null ? undefined : ((value['orderItems'] as Array<any>).map(OrderItemDTOToJSON)),
    };
}